implement max permroles

+++ src/constants.rs
@@ -44,4 +44,4 @@ pub const MAX_CHANNELS_PER_OVERRIDE: usize = 2;
 pub const MAX_OVERRIDES_PER_STARBOARD: i16 = 10; // todo
 
 // PermRole Validation
-pub const MAX_PERMROLES: i16 = 50; // todo
+pub const MAX_PERMROLES: i64 = 50;
